+++ src/service_identity/__init__.py
@@ -37,7 +37,7 @@ def __getattr__(name: str) -> str:
         "__url__": "",
         "__email__": "",
     }
-    if name not in dunder_to_metadata.keys():
+    if name not in dunder_to_metadata:
         raise AttributeError(f"module {__name__} has no attribute {name}")
 
     import warnings