+++ tests/test_packaging.py
@@ -1,14 +1,8 @@
-import sys
+from importlib import metadata
 
 import pytest
 
 import service_identity
-
-
-if sys.version_info < (3, 8):
-    import importlib_metadata as metadata
-else:
-    from importlib import metadata
 
 
 class TestLegacyMetadataHack: