+++ CHANGELOG.md
@@ -17,7 +17,7 @@ You can find out backwards-compatibility policy [here](https://github.com/pyca/s
 
 ### Changed
 
-- If a certificate doesn't contain any `subjectAltName`s, we now raise `service_identity.exceptions.CertificateError` instead of `service_identity.exceptions.VerificationError` to make the problem easier to debug.
+- If a certificate doesn't contain any `subjectAltName`s, we now raise `service_identity.CertificateError` instead of `service_identity.VerificationError` to make the problem easier to debug.
   [#67](https://github.com/pyca/service-identity/pull/67)