+++ CHANGELOG.md
@@ -13,7 +13,7 @@ You can find out backwards-compatibility policy [here](https://github.com/pyca/s
 <!-- changelog follows -->
 
 
-## [Unreleased](https://github.com/pyca/service-identity/compare/23.1.0...HEAD)
+## [24.1.0](https://github.com/pyca/service-identity/compare/23.1.0...24.1.0) - 2024-01-14
 
 ### Changed